+++ app/components/NotificationBanner/tests/index.test.js
@@ -1,10 +1,10 @@
-// import ErrorBanner from '../index';
+// import NotificationBanner from '../index';
 
 import expect from 'expect';
 // import { shallow } from 'enzyme';
 // import React from 'react';
 
-describe('<ErrorBanner />', () => {
+describe('<NotificationBanner />', () => {
   it('Expect to have unit tests specified', () => {
     expect(true).toEqual(false);
   });